tests/conftest.py
import yaml
import pytest
from pathlib import Path


FIXTURES_PATH = Path(__file__).parent.joinpath("fixtures/")


@pytest.fixture(scope="module")
def catalog_path():
    return FIXTURES_PATH.joinpath("catalog.yml")


@pytest.fixture(scope="module")
def catalog_data(catalog_path):
    with open(catalog_path, "r") as fh:
        return yaml.safe_load(fh)
